Attempt at strictly typing io module
This will most likely not work because the type system cannot be made to
understand a couple of things we're doing. This is especially apparent
with the overloads, where it seems that we get false errors, with claims
that types would be incompatible even though they work fine with our
dispatching system.